nfc_agent/requirements.txt
pyscard
requests
pystray
Pillow
pyinstaller
